server/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-speech
google-auth
google-auth-oauthlib
google-auth-httplib2
websockets
pydantic
python-docx
openpyxl
reportlab
python-pptx
gtts 
openai
fastapi-mail
whisper
python-multipart
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
aiohttp 
aiortc
pydub
fitz 
PyMuPDF
google-cloud-storage
tools
google-generativeai